Revert "Refactored e2e session cleaning and authentication."
This reverts commit 286269db48866bf7e5dd7409f15f27b08d0cfc90.
