+++ internal/testhelpers/e2e/clean.go
@@ -1,30 +1,52 @@
 package e2e
 
 import (
+	"os"
+	"testing"
+
+	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/pkg/platform/api/mono/mono_client/projects"
 	"github.com/ActiveState/cli/pkg/platform/api/mono/mono_client/users"
 	"github.com/ActiveState/cli/pkg/platform/api/mono/mono_models"
+	"github.com/ActiveState/cli/pkg/platform/authentication"
 )
 
-func (s *Session) cleanUser(username string) error {
-	projects, err := s.getProjects(username)
+func cleanUser(t *testing.T, username string, auth *authentication.Auth) error {
+	if os.Getenv(constants.APIHostEnvVarName) == "" {
+		err := os.Setenv(constants.APIHostEnvVarName, constants.DefaultAPIHost)
+		if err != nil {
+			return err
+		}
+		defer func() {
+			os.Unsetenv(constants.APIHostEnvVarName)
+		}()
+	}
+
+	err := auth.AuthenticateWithModel(&mono_models.Credentials{
+		Token: os.Getenv("PLATFORM_API_TOKEN"),
+	})
+	if err != nil {
+		return err
+	}
+
+	projects, err := getProjects(username)
 	if err != nil {
 		return err
 	}
 	for _, proj := range projects {
-		err = s.DeleteProject(username, proj.Name)
+		err = DeleteProject(username, proj.Name)
 		if err != nil {
 			return err
 		}
 	}
 
-	return s.deleteUser(username)
+	return deleteUser(username)
 }
 
-func (s *Session) getProjects(org string) ([]*mono_models.Project, error) {
+func getProjects(org string) ([]*mono_models.Project, error) {
 	params := projects.NewListProjectsParams()
 	params.SetOrganizationName(org)
-	listProjectsOK, err := s.auth.Client().Projects.ListProjects(params, s.auth.ClientAuth())
+	listProjectsOK, err := authentication.LegacyGet().Client().Projects.ListProjects(params, authentication.ClientAuth())
 	if err != nil {
 		return nil, err
 	}
@@ -32,16 +54,12 @@ func (s *Session) getProjects(org string) ([]*mono_models.Project, error) {
 	return listProjectsOK.Payload, nil
 }
 
-func (s *Session) DeleteProject(org, name string) error {
-	if s.auth == nil {
-		return nil // cannot do anything
-	}
-
+func DeleteProject(org, name string) error {
 	params := projects.NewDeleteProjectParams()
 	params.SetOrganizationName(org)
 	params.SetProjectName(name)
 
-	_, err := s.auth.Client().Projects.DeleteProject(params, s.auth.ClientAuth())
+	_, err := authentication.Client().Projects.DeleteProject(params, authentication.ClientAuth())
 	if err != nil {
 		return err
 	}
@@ -49,11 +67,11 @@ func (s *Session) DeleteProject(org, name string) error {
 	return nil
 }
 
-func (s *Session) deleteUser(name string) error {
+func deleteUser(name string) error {
 	params := users.NewDeleteUserParams()
 	params.SetUsername(name)
 
-	_, err := s.auth.Client().Users.DeleteUser(params, s.auth.ClientAuth())
+	_, err := authentication.Client().Users.DeleteUser(params, authentication.ClientAuth())
 	if err != nil {
 		return err
 	}
