+++ test/integration/push_int_test.go
@@ -72,7 +72,7 @@ func (suite *PushIntegrationTestSuite) TestInitAndPush() {
 	cp.ExpectLongString("Creating project")
 	cp.ExpectLongString("Project created")
 	cp.ExpectExitCode(0)
-	defer ts.DeleteProject(username, pname.String())
+	defer e2e.DeleteProject(username, pname.String())
 
 	// Check that languages were reset
 	pjfile, err := projectfile.Parse(pjfilepath)
@@ -154,7 +154,7 @@ func (suite *PushIntegrationTestSuite) TestPush_HeadlessConvert_NewProject() {
 	cp.SendLine(pname.String())
 	cp.Expect("Project created")
 	cp.ExpectExitCode(0)
-	defer ts.DeleteProject(username, pname.String())
+	defer e2e.DeleteProject(username, pname.String())
 
 	pjfile, err = projectfile.Parse(pjfilepath)
 	suite.Require().NoError(err)
@@ -205,7 +205,7 @@ func (suite *PushIntegrationTestSuite) TestPush_NoPermission_NewProject() {
 	cp.SendLine(pname.String())
 	cp.Expect("Project created")
 	cp.ExpectExitCode(0)
-	// Note: no need for ts.DeleteProject because newly created users and their projects are
+	// Note: no need for e2e.DeleteProject because newly created users and their projects are
 	// auto-cleaned by e2e.
 
 	pjfile, err = projectfile.Parse(pjfilepath)
@@ -263,7 +263,7 @@ func (suite *PushIntegrationTestSuite) TestCarlisle() {
 	cp.Send("y")
 	cp.Expect("Project created")
 	cp.ExpectExitCode(0)
-	ts.DeleteProject(username, pname.String())
+	e2e.DeleteProject(username, pname.String())
 }
 
 func (suite *PushIntegrationTestSuite) TestPush_Outdated() {
